# Conflicts:
#	src/main/java/org/example/bot/processor/Del/DelQuestionCommandProcessor.java
#	src/main/java/org/example/bot/processor/Del/DelQuestionProcessor.java
#	src/main/java/org/example/bot/processor/Del/DelTestProcessor.java
#	src/test/java/org/example/bot/service/QuestionProcessorTest.java
#	src/test/java/org/example/bot/service/TestProcessorTest.java

+++ src/main/java/org/example/bot/processor/Del/ConfirmDelQuestion.java
@@ -57,7 +57,7 @@ public class ConfirmDelQuestion extends AbstractCallbackProcessor {
         }
 
         QuestionEntity currentQuestion = optionalCurrentQuestion.get();
-        if (currentQuestion.getId() == Integer.parseInt(parts[1])) {
+        if (currentQuestion.getId() == Long.parseLong(parts[1])) {
             stateService.changeStateById(userId, UserState.DEFAULT);
             if (!parts[2].equals("YES")) {
                 return new BotResponse(String.format("Вопрос “%s” из теста “%s” не удален.",
